Format a date into a consistent human-readable format

Format a date into a consistent human-readable format

Usage

date_name(x, abbr_day = TRUE, abbr_month = TRUE, include_year = FALSE)

Arguments

x

a date in date format.

abbr_day

logical. Whether or not you would like the name of the day of the week to be abbreviated. Defaults to TRUE.

abbr_month

logical. Whether or not you would like the name of the month to be abbreviated. Defaults to TRUE.

include_year

logical. Whether you would like the date to include the year. Defaults to FALSE.

Value

Returns a brief commentary (including the figure for increases or decreases) as a string